fix error when resetting level and leveling rewards is not being used (i know it is a stupid mistake lmao)

+++ core/src/main/java/tk/bluetree242/discordsrvutils/systems/leveling/LevelingRewardsManager.java
@@ -155,6 +155,8 @@ public class LevelingRewardsManager {
 
     public void saveRewardCache() {
         try {
+            if (!rewardCacheFile.exists() && !needCache()) return;
+            else if (!rewardCacheFile.exists()) rewardCacheFile.createNewFile();
             FileWriter writer = new FileWriter(rewardCacheFile);
             writer.write(rewardCache.toString());
             writer.close();
